% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gopherdown.R
\docType{package}
\name{gopherdown}
\alias{gopherdown}
\title{gopherdown: A package for creating undergraduate, Masters, and PhD theses
using R Markdown}
\description{
This package creates theses/dissertations for the University of Minnesota.
}
\section{thesis_pdf}{

Creates an R Markdown thesis template as a PDF
}

