% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iflatex.R
\name{iflatex}
\alias{iflatex}
\title{Return latex control text if rendering latex}
\usage{
iflatex(txt)
}
\arguments{
\item{txt}{}
}
\value{
a string
}
\description{
Return latex control text if rendering latex
}
\examples{
\dontrun{
iflatex("\\\\vspace*{\\\\fill}")
}

}
