% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_caption.R
\name{render_caption}
\alias{render_caption}
\title{Render LaTeX captions}
\usage{
render_caption(caption, figname = "fig1", index = "index.Rmd", to = "latex")
}
\arguments{
\item{caption}{the markdown-formatted text you want to render as a figure
caption.}

\item{figname}{the name of the figure (as to not crowd the )}
}
\value{
a figure caption rendered in LaTeX
}
\description{
Captions are not rendered to latex by default. To aid in writing these, This
function will take in your markdown-formatted caption, and give you a latex
formatted caption.
}
\details{
This works by utilizing the \pkg{knitr} function
  \code{\link[knitr]{pandoc}}. It allows you to include references in your
  captions, which are not automatically rendered.
}
\examples{

# Setup for the example
# Note that this will already be set for you when you run the document
rootdir <- find.package("beaverdown")
rootdir <- paste0(rootdir, "/rmarkdown/templates/oregonstate/skeleton/")
knitr::opts_knit$set(root.dir = rootdir)


caption <- "This is some *text* to use as a caption [@angel2000]!"
# Make sure to set your index to whatever your project is called!
render_caption(caption, index = "skeleton.Rmd")

}
\seealso{
\code{\link{process_citations}}
}
